+++ README.md
@@ -56,10 +56,10 @@
 - [x] 플레이어는 게임을 시작할 때 배팅 금액을 정해야 한다.
     - [x] 베팅 금액이 1000원 미만이라면 예외처리
 - [x] 처음 두 장의 카드 합이 21일 경우 블랙잭
-- [ ] 플레이어가 블랙잭일 경우 베팅 금액의 1.5 배를 딜러에게 받는다.
-    - [ ] 딜러도 블랙잭일 경우에는 베팅한 금액을 돌려받는다.
-- [ ] 딜러가 21을 초과하면 플레이어들은 승리해 베팅 금액을 받는다.
-    - [ ] 플레이어도 21을 초과한 경우에는 베팅 금액을 잃는다.
+- [x] 플레이어가 블랙잭일 경우 베팅 금액의 1.5 배를 딜러에게 받는다.
+    - [x] 딜러도 블랙잭일 경우에는 베팅한 금액을 돌려받는다.
+- [x] 딜러가 21을 초과하면 플레이어들은 승리해 베팅 금액을 받는다.
+    - [x] 플레이어도 21을 초과한 경우에는 베팅 금액을 잃는다.
 
 ## 플레이어의 베팅금액이 10000원일 때의 경우
